requirements.txt
uvicorn
fastapi
uvicorn[standard]
gunicorn
pydantic
httpx
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-gcp-trace
google-auth
